+++ pyage/audio_backends/sound_wrapper.py
@@ -32,3 +32,17 @@ class SoundWrapper(ABC):
     @looping.setter
     def looping(self, looping: bool) -> None:
         raise NotImplementedError
+
+    @property
+    @abstractmethod
+    def position(self) -> float:
+        pass
+
+    @position.setter
+    def position(self, position: float) -> None:
+        raise NotImplementedError()
+
+    @property
+    @abstractmethod
+    def length(self) -> float:
+        pass