+++ pyage/audio_backends/synthizer/sound_wrapper.py
@@ -12,6 +12,7 @@ class SynthizerSoundWrapper(SoundWrapper):
     _context: synthizer.Context
     _generator: synthizer.BufferGenerator
     _last_position: float
+    _length: float
     _looping: bool
     _source: synthizer.Source3D
 
@@ -34,6 +35,7 @@ class SynthizerSoundWrapper(SoundWrapper):
         self._source.pause()
 
         self._last_position = 0.0
+        self._length = -1
         self._looping = False
 
     def play(self) -> None:
@@ -83,3 +85,21 @@ class SynthizerSoundWrapper(SoundWrapper):
     def looping(self, looping: bool) -> None:
         self._looping = looping
         self._generator.looping.value = looping
+
+    @property
+    def position(self) -> float:
+        self._last_position = self._generator.playback_position.value
+        return self._last_position
+
+    @position.setter
+    def position(self, position: float) -> None:
+        self._last_position = position
+        self._generator.playback_position.value = self._last_position
+
+    @property
+    def length(self) -> float:
+
+        if self._length < 0:
+            self._length = self._buffer.get_length_in_seconds()
+
+        return self._length